+++ armi/bookkeeping/db/tests/test_databaseInterface.py
@@ -24,7 +24,7 @@ from armi import __version__ as version
 from armi import interfaces
 from armi import runLog
 from armi import settings
-from armi.bookkeeping.db.database3 import Database3
+from armi.bookkeeping.db.database import Database
 from armi.bookkeeping.db.databaseInterface import DatabaseInterface
 from armi.cases import case
 from armi.context import PROJECT_ROOT
@@ -90,7 +90,7 @@ class TestDatabaseInterfaceBOL(unittest.TestCase):
 
             dbName = f"{self._testMethodName}.h5"
             self.dbi.initDB(fName=dbName)
-            self.db: Database3 = self.dbi.database
+            self.db: Database = self.dbi.database
             self.stateRetainer = self.r.retainState().__enter__()
             self.assertIsNotNone(self.dbi._db)
             self.dbi.interactBOL()
@@ -113,7 +113,7 @@ class TestDatabaseInterface(unittest.TestCase):
         )
         self.dbi = DatabaseInterface(self.r, self.o.cs)
         self.dbi.initDB(fName=self._testMethodName + ".h5")
-        self.db: Database3 = self.dbi.database
+        self.db: Database = self.dbi.database
         self.stateRetainer = self.r.retainState().__enter__()
 
     def tearDown(self):
@@ -174,7 +174,7 @@ class TestDatabaseInterface(unittest.TestCase):
         self.dbi.interactEOL()  # this also saves and closes db
 
         # reopen db to show EOL is written
-        with Database3(self._testMethodName + ".h5", "r") as db:
+        with Database(self._testMethodName + ".h5", "r") as db:
             self.assertTrue(db.hasTimeStep(r.p.cycle, r.p.timeNode, "EOL"))
             # and confirm that last time node is still there/separate
             self.assertTrue(db.hasTimeStep(r.p.cycle, r.p.timeNode))
@@ -217,7 +217,7 @@ class TestDatabaseInterface(unittest.TestCase):
             self.assertTrue(os.path.exists(self.dbi.database.fileName))
 
             # The copied file should have the newest time node
-            with Database3(self.dbi.database.fileName, "r") as db:
+            with Database(self.dbi.database.fileName, "r") as db:
                 for tn in range(timeNode + 1):
                     self.assertTrue(db.hasTimeStep(r.p.cycle, tn))
 
@@ -264,7 +264,7 @@ class TestDatabaseWriter(unittest.TestCase):
             :tests: R_ARMI_DB_QA
         """
         with h5py.File("test_writeSystemAttributes.h5", "w") as h5:
-            Database3.writeSystemAttributes(h5)
+            Database.writeSystemAttributes(h5)
 
         with h5py.File("test_writeSystemAttributes.h5", "r") as h5:
             self.assertIn("user", h5.attrs)
@@ -457,7 +457,7 @@ class TestDatabaseReading(unittest.TestCase):
         self.assertEqual(len(r.core.blocksByName), 95)
 
     def test_growToFullCore(self):
-        with Database3(self.dbName, "r") as db:
+        with Database(self.dbName, "r") as db:
             r = db.load(0, 0, allowMissing=True)
 
         # test partial core values
@@ -471,7 +471,7 @@ class TestDatabaseReading(unittest.TestCase):
         self._fullCoreSizeChecker(r)
 
     def test_growToFullCoreWithCS(self):
-        with Database3(self.dbName, "r") as db:
+        with Database(self.dbName, "r") as db:
             r = db.load(0, 0, allowMissing=True)
 
         r.core.growToFullCore(self.cs)
@@ -498,7 +498,7 @@ class TestDatabaseReading(unittest.TestCase):
         self._fullCoreSizeChecker(r)
 
     def test_readWritten(self):
-        with Database3(self.dbName, "r") as db:
+        with Database(self.dbName, "r") as db:
             r2 = db.load(0, 0, self.cs)
 
         for a1, a2 in zip(self.r.core, r2.core):
@@ -539,7 +539,7 @@ class TestDatabaseReading(unittest.TestCase):
             )
 
     def test_readWithoutInputs(self):
-        with Database3(self.dbName, "r") as db:
+        with Database(self.dbName, "r") as db:
             r2 = db.load(0, 0)
 
         for b1, b2 in zip(self.r.core.getBlocks(), r2.core.getBlocks()):
@@ -550,7 +550,7 @@ class TestDatabaseReading(unittest.TestCase):
             assert_allclose(b.p.flux, 1e6 * bi)
 
     def test_variousTypesWork(self):
-        with Database3(self.dbName, "r") as db:
+        with Database(self.dbName, "r") as db:
             r2 = db.load(1, 1)
 
         b1 = self.r.core.getFirstBlock(Flags.FUEL)
@@ -577,7 +577,7 @@ class TestDatabaseReading(unittest.TestCase):
         assert_allclose(numDensVec1, numDensVec2)
 
     def test_timesteps(self):
-        with Database3(self.dbName, "r") as db:
+        with Database(self.dbName, "r") as db:
             # build time steps in the DB file
             timesteps = []
             for cycle in range(self.nCycles):
